[add]それぞれのworkcomponentを押すと、作品詳細画面に遷移する機能の追加
