+++ src/reducers/Works/WorksState.js
@@ -1,28 +1,26 @@
-import { GET_WORKS_SUCCESS, POST_WORKS_SUCCESS, POST_WORKS_DATA_SUCCESS} from "../../actions/Works/ActionType";
+import { GET_WORKS_SUCCESS, POST_WORKS_SUCCESS, POST_WORKS_DATA_SUCCESS } from "../../actions/Works/ActionType";
 
-const initialState = {
-  works: {}
-};
+const initialState = {};
 
 
 const WorksState = (state = initialState, action) => {
-    switch (action.type) {
-        case GET_WORKS_SUCCESS:
-          return {
-            works:{...action.works}
-          }
-        case POST_WORKS_SUCCESS:
-          return {
-            works:{...action.works}
-          }
-        case POST_WORKS_DATA_SUCCESS:
-        console.log(action.worksData);
-          return {
-            works:{...action.works}
-          }
-        default:
-          return state;
-    }
+  switch (action.type) {
+    case GET_WORKS_SUCCESS:
+      return {
+        ...action.works
+      }
+    case POST_WORKS_SUCCESS:
+      return {
+        ...action.works
+      }
+    case POST_WORKS_DATA_SUCCESS:
+      console.log(action.worksData);
+      return {
+        ...action.works
+      }
+    default:
+      return state;
+  }
 }
 
 export default WorksState
